Use DBus directly for Unity protocol
The libunity library detects if the program is inside a snap,
in which case it modifies the passed .desktop file to match the
expected one in a snapped program. Unfortunately, this prevents
us to pose as other programs to put a progress bar, so we need
to work directly with the DBus protocol itself.